requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dateutil==2.9.0
numpy==2.1.2
pandas==2.2.3
python-multipart==0.0.12
sqlacodegen==3.0.0rc5
supertokens-python==0.29.1
python-dotenv==1.0.1

# Testing dependencies
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0
httpx==0.27.2
faker==30.3.0
